Create RoutingTableIncrementalDiff non-remote entity.
